compose: Move initialize to new compose_setup module.
