+++ tools/test-js-with-node
@@ -69,6 +69,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_fade.js",
         "web/src/compose_popovers.js",
         "web/src/compose_recipient.js",
+        "web/src/compose_setup.js",
         "web/src/compose_state.js",
         "web/src/compose_textarea.ts",
         "web/src/compose_tooltips.js",
