+++ web/tests/compose_actions.test.js
@@ -64,7 +64,7 @@ mock_esm("../src/resize", {
 
 const people = zrequire("people");
 
-const compose = zrequire("compose");
+const compose_setup = zrequire("compose_setup");
 const compose_state = zrequire("compose_state");
 const compose_actions = zrequire("compose_actions");
 const message_lists = zrequire("message_lists");
@@ -224,11 +224,11 @@ test("start", ({override, override_rewire, mock_template}) => {
     };
 
     let abort_xhr_called = false;
-    override_rewire(compose, "abort_xhr", () => {
+    override_rewire(compose_setup, "abort_xhr", () => {
         abort_xhr_called = true;
     });
 
-    compose_actions.register_compose_cancel_hook(compose.abort_xhr);
+    compose_actions.register_compose_cancel_hook(compose_setup.abort_xhr);
     $("#compose-textarea").set_height(50);
 
     assert_hidden("#compose_controls");
